poc/requirements.txt
# Minimal POC Requirements

# API clients (optional - only needed if using cloud APIs)
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Core dependencies (required)
httpx>=0.25.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
rich>=13.0.0
python-dotenv>=1.0.0
